+++ src/core/misc/random.hpp
@@ -9,7 +9,7 @@ namespace koishi
 namespace core
 {
 PolyFunction( DRand48, Require<Host> )(
-  ()->double {
+  ()->float {
 	  static unsigned long long seed = ( ( (long long int)time( nullptr ) ) << 16 ) | ::rand();
 
 	  constexpr auto m = 0x100000000LL;
@@ -17,11 +17,11 @@ PolyFunction( DRand48, Require<Host> )(
 	  constexpr auto a = 0x5DEECE66DLL;
 	  seed = ( a * seed + c ) & 0xFFFFFFFFFFFFLL;
 	  unsigned int x = seed >> 16;
-	  return ( (double)x / (double)m );
+	  return ( (float)x / (float)m );
   } );
 
 PolyFunction( FakeRand, Require<Device>, Require<Host> )(
-  ()->double {
+  ()->float {
 	  return 0.5;
   } );
